+++ web/CHANGELOG.md
@@ -8,7 +8,7 @@
 - Added `URL.toDart` and `Uri.toJS` extension methods.
 - Added missing `Document` and `Window` pointer event getters: `onDrag*`,
   `onTouch*`, `onMouse*`.
-- Added `JSLiveNodeListWrapper` to support mutable operations on nodes lists. 
+- Added `JSLiveNodeListWrapper` to support mutable operations on node lists. 
 - Added `childNodesAsList` to `Node` and `childrenAsList` to `Element` via 
   extensions.